Кerberos support (#177)
merging into kerberos-support, will raise a separate PR to master from here. If the original submitters of this PR could raise a new PR against the kerberos-support branch addressing some of the comments in this PR that would be great. We'll try keep everyone who has commented on this in the loop for the new PR and would really appreciate it if you could help us with testing the new branch when it's closer to being ready to do as we don't currently have a Kerberised environment we can use for testing.